requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-multipart
python-dotenv
httpx
